Finished Excercise 4 Chapter 12
There's still a bug that with the new spreadsheet nothing shows in LibreOffice before changing column and row sizes, but I don't think I'll be adding column width and row height re-flowing to this script. Hrm, maybe, but I'll need to figure out how to calculate sane values for this or whether there is a built-in function to do it.